feat: historical data
