+++ src/shared/lib/date/index.ts
@@ -47,3 +47,14 @@ export const addDays = (date: Date, days: number) => {
 
     return new Date(date.valueOf() + millisecondsToAdd);
 };
+
+export const formatDate = (ts: number) => {
+    const date = new Date(ts);
+
+    const hours = date.getHours().toString().padStart(2, '0');
+    const minutes = date.getMinutes().toString().padStart(2, '0');
+    const month = (date.getMonth() + 1).toString().padStart(2, '0');
+    const day = date.getDate().toString().padStart(2, '0');
+
+    return `${date.getFullYear()}-${month}-${day} ${hours}:${minutes}`;
+};
